Set up a few routes using routers
